Issue 'too many matches' warning for DirAssets also
